fix(params): accept Quantity values; centralize _to_hashable in _params
The previous commit fixed mechanism_signature in _compute/runtime.py,
but the same root cause (Quantity unhashable in newer saiunit) also
trips Params.__init__'s explicit hashability check and Params.__hash__:

  TypeError: Params value for 'g_max' must be hashable
             (arrays are rejected); got 'Quantity'.

Move the _to_hashable helper from _compute/runtime.py into
mech/_params.py (its more natural home — Params already encodes
"hashable mapping" semantics) and use it in both:

- Params.__init__ validation accepts Quantity (and other compound
  values containing Quantity) but still rejects raw numpy/jax arrays
- Params.__hash__ feeds _to_hashable(value) into the frozenset
- mechanism_signature continues to use the same helper, now imported
  from _params

Order-insensitive equality/hashing of Params and dedup semantics for
mechanism layouts are preserved.
